RES-2479: Move nupic.research.frameworks.sigopt to packages
