+++ nupic/research/frameworks/vernon/trainables.py
@@ -35,7 +35,6 @@ from ray.tune.resources import Resources
 from ray.tune.result import DONE, RESULT_DUPLICATE
 from ray.tune.utils import warn_if_slow
 
-from nupic.research.frameworks.sigopt import SigOptExperiment
 from nupic.research.frameworks.vernon.experiment_utils import get_free_port
 
 os.environ["HDF5_USE_FILE_LOCKING"] = "FALSE"
@@ -446,96 +445,6 @@ class DistributedTrainable(RemoteTrainableBase):
         except Exception:
             self._kill_workers()
             raise
-
-
-class SigOptTrainableMixin:
-    """
-    This class updates the config using SigOpt before the models and workers are
-    instantiated, and updates the result using SigOpt once training completes.
-    """
-
-    def _process_config(self, config):
-        """
-        :param config:
-            Dictionary configuration of the trainable
-
-            - sigopt_experiment_id: id of experiment
-            - sigopt_config: dict to specify configuration of sigopt experiment
-            - sigopt_experiment_class: class inherited from `SigoptExperiment` which
-                                       characterizes how the trainable will get and
-                                       utilize suggestions
-
-        """
-        # Update the config through SigOpt.
-        self.sigopt = None
-        if "sigopt_config" in config:
-            assert config.get("sigopt_experiment_id", None) is not None
-
-            # Check for user specified sigopt-experiment class.
-            experiment_class = config.get(
-                "sigopt_experiment_class", SigOptExperiment)
-            assert issubclass(experiment_class, SigOptExperiment)
-
-            # Instantiate experiment.
-            self.sigopt = experiment_class(
-                experiment_id=config["sigopt_experiment_id"],
-                sigopt_config=config["sigopt_config"])
-
-            self.logger.info(
-                f"Sigopt execution order: {pformat(self.sigopt.get_execution_order())}")
-
-            # Get suggestion and update config.
-            self.suggestion = self.sigopt.get_next_suggestion()
-            self.sigopt.update_config_with_suggestion(config, self.suggestion)
-            print("SigOpt suggestion: ", self.suggestion)
-            print("Config after Sigopt:")
-            pprint(config)
-            self.epochs = config["epochs"]
-
-            # Get names of performance metrics.
-            assert "metrics" in config["sigopt_config"]
-            self.metric_names = [
-                metric["name"] for metric in config["sigopt_config"]["metrics"]
-            ]
-            assert len(self.metric_names) > 0, \
-                "For now, we only update the observation if a metric is present."
-
-    def _process_result(self, result):
-        """
-        Update sigopt with the new result once we're at the end of training.
-        """
-
-        super()._process_result(result)
-
-        if self.sigopt is not None:
-            result["early_stop"] = result.get("early_stop", 0.0)
-            if self.iteration >= self.epochs - 1:
-                result["early_stop"] = 1.0
-                # check that all metrics are present
-                print(result)
-                for name in self.metric_names:
-                    if result[name] is not None:
-                        self.logger.info(f"Updating observation {name} with value=",
-                                         result[name])
-                    else:
-                        self.logger.warning(f"No value: {name}")
-
-                # Collect and report relevant metrics.
-                values = [
-                    dict(name=name, value=result[name])
-                    for name in self.metric_names
-                ]
-                self.sigopt.update_observation(self.suggestion, values=values)
-                print("Full results: ")
-                pprint(result)
-
-
-class SigOptRemoteProcessTrainable(SigOptTrainableMixin, RemoteProcessTrainable):
-    pass
-
-
-class SigOptDistributedTrainable(SigOptTrainableMixin, DistributedTrainable):
-    pass
 
 
 class DebugTrainable(Trainable):
