+++ packages/sigopt/nupic/research/frameworks/sigopt/mixins/one_cycle_lr_params.py
