+++ packages/sigopt/nupic/research/frameworks/sigopt/mixins/sgd_params.py
