+++ packages/sigopt/setup.cfg
@@ -20,7 +20,7 @@
 #
 # ------------------------------------------------------------------------------
 [metadata]
-name = nupic.research.archive
+name = nupic.research.sigopt
 version = 0.0.1.dev0
 
 [options]
@@ -28,11 +28,8 @@ zip_safe = False
 packages = find_namespace:
 install_requires =
     nupic.research
+    sigopt
 
 [options.packages.find]
 include = nupic.research.*
-
-[options.extras_require]
-sigopt =
-    sigopt
