+++ packages/sigopt/setup.py
@@ -19,20 +19,6 @@
 #  http://numenta.org/licenses/
 #
 # ------------------------------------------------------------------------------
-[metadata]
-name = nupic.research.archive
-version = 0.0.1.dev0
+from setuptools import setup
 
-[options]
-zip_safe = False
-packages = find_namespace:
-install_requires =
-    nupic.research
-
-[options.packages.find]
-include = nupic.research.*
-
-[options.extras_require]
-sigopt =
-    sigopt
-
+setup()
