+++ setup.cfg
@@ -65,7 +65,6 @@ install_requires =
     ray[tune]~=0.8.3
     seaborn
     wandb
-    sigopt
 
 [options.extras_require]
 # Project specific dependencies
@@ -83,7 +82,6 @@ dendrites =
     ptitprince==0.2.5
     seaborn==0.11.1
     sklearn
-    nupic.research
 
 transformers =
     ray[tune]==1.4.0
